change default admin to False and fix travis build
