--- app/.travis.yml
@@ -1,16 +1,0 @@
-language: python
-python:
-  - "3.7"
-
-# command to install dependencies
-install:
-  - pip install -r requirements.txt
-  - pip install nose
-
-# command to run tests
-script:
- - nosetests --with-coverage
-
-
-after_success:
-  - coveralls
